Revert "Merge remote-tracking branch 'upstream/master'"
This reverts commit fccfa1744a5dcb9bb0a63cae1300637bf83806ca, reversing
changes made to 634c83e4342e8a4ac75b35f2a5ea5fda05d2f878.

+++ modules/juce_core/native/juce_android_SystemStats.cpp
@@ -229,7 +229,7 @@ int64 Time::getHighResolutionTicksPerSecond() noexcept
 
 double Time::getMillisecondCounterHiRes() noexcept
 {
-    return (double) getHighResolutionTicks() * 0.001;
+    return getHighResolutionTicks() * 0.001;
 }
 
 bool Time::setSystemTimeToThisTime() const
